Fix _open_emacs word splitting behavior (lack thereof)
Fish does not do word splitting; instead, every variable is an array,
each element of which gets passed as separate arguments when expanded.
Therefore, to fix the error I simply pushed --create-frame onto the end
and interpolated the same way.
Another solution ould have been to delete the double quotes in the
original code, but I preferred the lack of repetition in the new implementation.