Change compute-tangle-file to global function
Issue was that when babel executed the source block on startup, it would
execute the block in another buffer, meaning that the local binding
would not persist into the actual buffer.
To remedy this, we just put it in global scope.
Alternatively, we could have used `eval-last-sexp` with the source block
to have Emacs execute it normally, but this would have been a little bit janky.